refactor: move writer to internal
